+++ src/index.ts
@@ -1,6 +1,7 @@
 export * from '@nestjs/swagger';
 
+export { SwaggerModule } from './module';
+
 export * from './decorator';
-export * from './module';
 export * from './provider';
 export * from './struct';