+++ CHANGELOG.md
@@ -1,3 +1,39 @@
+## [8.1.4](https://github.com/vodyani/swagger/compare/v8.1.3...v8.1.4) (2022-09-07)
+
+
+### Bug Fixes
+
+* deps ([3fdcecf](https://github.com/vodyani/swagger/commit/3fdcecfbedfdcaf6887ce84aeede7ccde22cd5ce))
+
+## [8.1.3](https://github.com/vodyani/swagger/compare/v8.1.2...v8.1.3) (2022-08-31)
+
+
+### Bug Fixes
+
+* add extra model ([00132e3](https://github.com/vodyani/swagger/commit/00132e3a783c36a53ec2be0366aa96d133d8fb08))
+
+## [8.1.2](https://github.com/vodyani/swagger/compare/v8.1.1...v8.1.2) (2022-08-31)
+
+
+### Bug Fixes
+
+* remove invalid type ([673fe55](https://github.com/vodyani/swagger/commit/673fe5542cbb0b8d88aa79b7a00f125967549bbd))
+
+## [8.1.1](https://github.com/vodyani/swagger/compare/v8.1.0...v8.1.1) (2022-08-31)
+
+
+### Bug Fixes
+
+* add form data api decorator ([bef2bf2](https://github.com/vodyani/swagger/commit/bef2bf2baea4b23aee3e664f00b212dbec6013b2))
+
+# [8.1.0](https://github.com/vodyani/swagger/compare/v8.0.1...v8.1.0) (2022-08-30)
+
+
+### Features
+
+* add some decorators ([db41441](https://github.com/vodyani/swagger/commit/db414419ca2efdd30cf73d5ba993d5d584ff8b3b))
+* integrate dynamic loading module and reconstruct working mode ([c4cbb7a](https://github.com/vodyani/swagger/commit/c4cbb7ac19bd1edb199ae8fcac301b018e154271))
+
 # [8.1.0-beta.1](https://github.com/vodyani/swagger/compare/v8.0.1...v8.1.0-beta.1) (2022-08-30)